chatbot/chatbotbackend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
google-generativeai==0.3.2
chromadb==0.4.22
sentence-transformers==2.3.1
pydantic==2.5.3
python-jose[cryptography]==3.3.0
numpy<2.0.0
